Move weather location type to weather module
